Adds basic_units, plot_hamiltonian, bug fixes TODO
basic_units.py is a file downloaded from the matplotlib website that
should allow for easy plotting of radian units. DirectionModel now has
plot_weighting and plot_hamiltonian, and there have been tweaks to
various parts of the code. plot_targets.py is now being used to debug,
which is definitely needed given the result of get_direction in the
current version.