Bug fixes for point targets, argmax. Adds track plotting
Point targets had an error related to periodicity in the theta_mesh that
threw an error on running and was fixed. Argmax had a much worse error
related to floating point error deciding the direction in a way that
would be consistent for a given theta mesh resolution. It resulted in
random walkers always going toward the same point mass target. Finally,
a flag has been added to random walker plotting to turn on and off track
plotting.